backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Core ML libraries with compatible versions
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
missingno==0.5.2
kaleido==0.2.1

# ML frameworks
lightgbm==4.0.0
xgboost==1.7.6
flaml==2.1.1

# Explainability
shap==0.43.0
lime==0.2.0.1

# Database and storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
minio==7.2.0

# Monitoring
prometheus-client==0.19.0

# AI/ML extras
google-generativeai==0.3.2
Pillow==10.1.0
opencv-python==4.8.1.78

# Async and messaging
celery==5.3.4
redis==5.0.1
websockets==12.0

# Optional - only install if needed
# torch==2.1.1
# torchvision==0.16.1
# transformers==4.36.2
# sentence-transformers==2.2.2
# nltk==3.8.1
# spacy==3.7.2
# statsmodels==0.14.0
# prophet==1.1.5
# tsfresh==0.20.1
# presidio-analyzer==2.2.33
# presidio-anonymizer==2.2.33
# optuna==3.4.0
# openpyxl==3.1.2
# xlrd==2.0.1
